# app/openapi.yml
openapi: 3.0.3
info:
  title: WMO / Zorgned koppel api
  description: Dit is het koppelstuk tussen Mijn Amsterdam en WMO / Zorgned
  version: 2.0.0
  contact:
    email: datapunt@amsterdam.nl
    name: Datapunt Support
    url: https://data.amsterdam.nl/
paths:
  /wmoned/voorzieningen:
    parameters:
      - name: x-saml-attribute-token1
        in: header
        description: TMA encoded BSN
        required: true
        schema:
          type: string
    get:
      summary: Returns a list of actual voorzieningen
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: "#/components/schemas/Voorzieningen"
                  status:
                    type: string
                    enum:
                      - "OK"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /status/health:
    get:
      summary: returns "OK"
      responses:
        "200":
          description: Returns "OK"
          content:
            application/json:
              schema:
                type: string
                enum:
                  - "OK"
components:
  schemas:
    Voorzieningen:
      type: array
      items:
        $ref: "#/components/schemas/Voorziening"
    Voorziening:
      type: object
      properties:
        title:
          type: string
        dateStart:
          description: The date the request for this product was started
          type: string
          format: date
          example: "2020-06-16"
        dateEnd:
          description: The date the request for this product was finished
          type: string
          format: date
          example: "2020-06-16"
          nullable: true
        supplier:
          type: string
          description: The name of the company which supplies the product
          nullable: true
        isActual:
          description: Is the product delivery actual
          type: boolean
        dateDecision:
          description: The date a decision was made about the request for this product
          type: string
          format: date
          example: "2020-06-16"
          nullable: true
        itemTypeCode:
          type: string # Enum?
        deliveryType:
          type: string # Enum?
        serviceDateStart:
          description: The date on which the product started/delivered
          type: string
          format: date
          example: "2020-06-16"
          nullable: true
        serviceDateEnd:
          description: The date on which the product ended/revoked
          type: string
          format: date
          example: "2020-06-16"
          nullable: true
        serviceOrderDate:
          description: The date on which the request was made to start delivering the product
          type: string
          format: date
          example: "2020-06-16"
          nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - "ERROR"
        message:
          type: string